Updated unittest and Problems Classmethod
