+++ openqaoa/problems/knapsack.py
@@ -16,7 +16,7 @@ import itertools
 
 import numpy as np
 
-from .helper_functions import check_kwargs
+from ..utilities import check_kwargs
 from .problem import Problem
 from .qubo import QUBO
