+++ openqaoa/problems/minimumvertexcover.py
@@ -16,7 +16,7 @@ from collections import Counter
 
 import networkx as nx
 
-from .helper_functions import check_kwargs
+from ..utilities import check_kwargs
 from .problem import Problem
 from .qubo import QUBO
