+++ openqaoa/utilities.py
@@ -27,9 +27,6 @@ import networkx as nx
 
 from .qaoa_parameters import Hamiltonian, PauliOp, QAOAVariationalBaseParams
 from .qaoa_parameters.gatemap import *
-from .problems import (Knapsack, SlackFreeKnapsack, MaximumCut, 
-                       MinimumVertexCover, NumberPartition, ShortestPath, TSP, 
-                       QUBO)
 
 
 def X_mixer_hamiltonian(n_qubits: int,
@@ -1459,57 +1456,45 @@ def is_valid_uuid(uuid_to_test: str) -> bool:
         # If it's a value error, then the string is not a valid string for a UUID.
         return False
 
-
 ################################################################################
-# PROBLEMS CLASS UTILITY
+# CHECKING FUNCTION
 ################################################################################
-
-def create_problem_from_dict(problem_dict: dict) -> Problem:
+    
+def check_kwargs(list_expected_params, list_default_values, **kwargs):
     """
-    Creates an object of the class corresponding to the problem type in the input instance, with the same attributes as the input instance.
-    Parameters
-    ----------
-    problem_instance: dict
-        The input instance.
-    Returns
-    -------
-        An object of the class corresponding to the problem type in the input instance.
+    Checks that the given list of expected parameters can be found in the
+    kwargs given as input. If so, it returns the parameters from kwargs, else
+    it raises an exception.
+
+    Args:
+        list_expected_params: List[str]
+            List of string containing the name of the expected parameters in
+            kwargs
+        list_default_values: List
+            List containing the deafult values of the expected parameters in
+            kwargs
+        **kwargs:
+            Keyword arguments where keys are supposed to be the expected params
+
+    Returns:
+        A tuple with the actual expected parameters if they are found in kwargs.
+
+    Raises:
+        ValueError:
+            If one of the expected arguments is not found in kwargs and its
+            default value is not specified.
     """
 
-    # copy the instance to avoid modifying the original instance
-    problem_instance = problem_instance.copy()
+    def check_kwarg(expected_param, default_value, **kwargs):
+        param = kwargs.pop(expected_param, default_value)
 
-    # mapper from problem type to class
-    problem_mapper = { 
-        "generic_qubo": QUBO,
-        "tsp": TSP,
-        "number_partition": NumberPartition,
-        "maximum_cut": MaximumCut,
-        "knapsack": Knapsack,
-        "slack_free_knapsack": SlackFreeKnapsack,
-        "minimum_vertex_cover": MinimumVertexCover,
-        "shortest_path": ShortestPath,
-    }
+        if param is None:
+            raise ValueError(f"Parameter '{expected_param}' should be specified")
 
-    # check if the problem type is in the mapper
-    assert problem_instance["problem_type"] in problem_mapper, f"Problem type {problem_instance['problem_type']} not supported."
+        return param
 
-    # get the class corresponding to the problem type
-    problem_class = problem_mapper[problem_instance.pop('problem_type', "generic_qubo")]
+    params = []
+    for expected_param, default_value in zip(list_expected_params, list_default_values):
+        params.append(check_kwarg(expected_param, default_value, **kwargs))
 
-    # check if the problem type is QUBO, if so, raise an exception
-    if problem_class is QUBO:
-        raise Exception("This method does not work for generic QUBO. The input instance has type `generic_qubo`. You can use the `from_dict` method of the `QUBO` class instead.")
-
-    # if the instance has a graph, convert it to a networkx graph
-    if 'G' in problem_instance:
-        problem_instance['G'] = nx.node_link_graph(problem_instance['G'])
-
-    # erase the keys that are not arguments of the class
-    arguments = inspect.getfullargspec(problem_class).args[1:]
-    for key in problem_instance.copy():
-        if key not in arguments:
-            del problem_instance[key]
-
-    # create the problem instance and return it
-    return problem_class(**problem_instance)
+    return tuple(params)
